+++ src/component/AleBoxView/index.tsx
@@ -7,65 +7,20 @@ import { useNavigation } from "@react-navigation/native";
 
 
 //props = {data:[]}
-function Box(){
+function Box({data}){
     const navigation=useNavigation()
-
-
-    const team = [
-        {
-            avatar:require("../Images/Avatar-base(5).png"),
-            firstName:"Humberto ",
-            lastName:"Alonso",
-            area:"UX&UI Manager"
-        },
-        {
-            avatar:require("../Images/Avatar-base(4).png"),
-            firstName:"Luis ",
-            lastName:"Mota",
-            area:"Trainne UX&UI "
-        },
-        {
-            avatar:require("../Images/Avatar-base(3).png"),
-            firstName:"César",
-            lastName:"Ramón",
-            area:"Trainne UX&UI "
-        },
-        {
-            avatar:require("../Images/Avatar-base(2).png"),
-            firstName:"Aline",
-            lastName:"Palacios",
-            area:"Trainne UX&UI "
-        },
-        {
-            avatar:require("../Images/Avatar-base(1).png"),
-            firstName:"Nathali",
-            lastName:"Ortiz",
-            area:"Trainne UX&UI "
-        },
-        {
-            avatar:require("../Images//Avatar-base.png"),
-            firstName:"Vanessa",
-            lastName:"Luna",
-            area:"Trainne UX&UI "
-        },
-        {
-            avatar:require("../Images//Avatar-base.png"),
-            firstName:"Vanessa",
-            lastName:"Luna",
-            area:"Trainne UX&UI "
-        },
-    ]
     return(
        <SafeAreaView>
  
     {        
        //map me genera un clone del arreglo y lo podemos modificar
         //array.map( (item,index,array,arrayOriginal)=>)
-        //[]    //{}
-        team.map((item,index)=>{
-            return(
+        //[]    //{ITEM={AVATAR,NAME,LAST,AREA}
+        data.map((item,index)=>
+        {
+            return(     
 
-                <TouchableOpacity onPress={()=>navigation.navigate("Perfil")}>
+                <TouchableOpacity key={index} onPress={()=>navigation.navigate("Perfil")}>
                     <View style={styles.squareShape}>
                         <View style={{alignItems:"center", flexDirection:"row"}}>
                             <Image style={styles.fotostyle} source={item.avatar}/>
@@ -77,9 +32,8 @@ function Box(){
                         <PenIcon/>
                     </View>
                  </TouchableOpacity>
-            )
+                )
         })
-
     }
        </SafeAreaView>
     )
