+++ src/scenes/organization/index.tsx
@@ -1,4 +1,4 @@
-import React, {useMemo,useRef,useState} from "react";
+import React, {useMemo,useRef,useState,useInsertionEffect} from "react";
 import { SafeAreaView, View, Text, Image, StyleSheet, ScrollView, Button, Alert, TouchableOpacity} from "react-native";
 import Input from "../../component/AleBoxView/input";
 import Box from "../../component/AleBoxView";
@@ -12,6 +12,50 @@ import { useNavigation } from "@react-navigation/native";
 import BottomSheet, {BottomSheetBackdrop} from "@gorhom/bottom-sheet";
 import FilterSheet from "../../component/BottomSheets";
 
+const team = [
+    {
+        avatar:require("../../component/Images/Avatar-base(5).png"),
+        firstName:"Humberto ",
+        lastName:"Alonso",
+        area:"UX&UI Manager"
+    },
+    {
+        avatar:require("../../component/Images/Avatar-base(4).png"),
+        firstName:"Luis ",
+        lastName:"Mota",
+        area:"Trainne UX&UI "
+    },
+    {
+        avatar:require("../../component/Images/Avatar-base(3).png"),
+        firstName:"César",
+        lastName:"Ramón",
+        area:"Trainne UX&UI "
+    },
+    {
+        avatar:require("../../component/Images/Avatar-base(2).png"),
+        firstName:"Aline",
+        lastName:"Palacios",
+        area:"Trainne UX&UI "
+    },
+    {
+        avatar:require("../../component/Images/Avatar-base(1).png"),
+        firstName:"Nathali",
+        lastName:"Ortiz",
+        area:"Trainne UX&UI "
+    },
+    {
+        avatar:require("../../component/Images//Avatar-base.png"),
+        firstName:"Vanessa",
+        lastName:"Luna",
+        area:"Trainne UX&UI "
+    },
+    {
+        avatar:require("../../component/Images//Avatar-base.png"),
+        firstName:"Vanessa",
+        lastName:"Luna",
+        area:"Trainne UX&UI "
+    },
+]
 
 function OrganScene(){
   const navigation = useNavigation()
@@ -19,6 +63,9 @@ function OrganScene(){
   const bottomSheetRef = useRef<BottomSheet>(null);
  // variables
  const snapPoints = useMemo(() => ['25%', '80%'], []);
+ //hook
+ const [searchText,setSearchText] = useState('')
+ //console.log("search",team.filter((item)=> item.firstName.toLowerCase().includes(searchText.toLowerCase()) ))
   
  return (
       
@@ -33,8 +80,7 @@ function OrganScene(){
                    </View>
               </View>
           </View>
-           
-
+          {/*SearchBar*/}
           <View style={styles.buttonSearch}>
                <View style={{
                    flexDirection:"row",
@@ -47,12 +93,15 @@ function OrganScene(){
                    alignItems:"center"
                }}>
                    <SearchIcon/>
-                   <Input style={{marginLeft:12}}
+                   <Input 
+                   onChangeText={(text)=> setSearchText(text)}
+                   style={{marginLeft:12}}
                    placeholderTextColor={"#333"}
                    placeholder="Buscar en Mi empresa"/>
                </View>
               
           </View>
+          {/*Buttons Filter*/}
           <View style={styles.filter}>
               <View style={styles.buttonShape}><Text style={styles.textFiltwo}>Filtros</Text></View>   
               <TouchableOpacity onPress={()=>bottomSheetRef.current?.expand(BottomSheet)} style={styles.buttonShapeFour}><FilterIcon/></TouchableOpacity>
@@ -62,11 +111,15 @@ function OrganScene(){
               </TouchableOpacity>
               
           </View>
-
+          
+          
+         {/*Content*/}
           <View style={styles.main}>
           <ScrollView>
               <Text style={styles.texthree}> Departamento de Diseño </Text>
-              <Box/>
+              <Box data={
+                team.filter((item)=> item.firstName.toLowerCase().includes(searchText.toLowerCase()) || item.lastName.toLowerCase().includes(searchText.toLowerCase()) )
+              }/>
           </ScrollView>
           </View>
